Add ALLOW_SAME_WINDOW flag for content-script <-> page messaging
When CONFIG.ALLOW_SAME_WINDOW is true, allow messages to the the same window and ignore requests from this instance. Turned off by default for compatibility.

The idea is that you have another instance in another context/sandbox bound to the same window that wants to talk to us. This is common in WebExtensions.

This is similar to CONFIG.ALLOW_SAME_ORIGIN but also prevents post-robot from confusion when it sees its own requests by discarding them early.